+++ packages/core/src/interpreter.ts
@@ -24,6 +24,7 @@ import type {
   PersistedStateFrom,
   SnapshotFrom,
   AnyActorRef,
+  OutputFrom,
   DoneActorEvent
 } from './types.ts';
 import {
@@ -35,6 +36,7 @@ import {
   Subscription
 } from './types.ts';
 import { toObserver } from './utils.ts';
+import { TlsOptions } from 'tls';
 
 export type SnapshotListener<TLogic extends AnyActorLogic> = (
   state: SnapshotFrom<TLogic>
@@ -76,20 +78,13 @@ const defaultOptions = {
   devTools: false
 };
 
-type InternalStateFrom<TLogic extends ActorLogic<any, any, any>> =
-  TLogic extends ActorLogic<infer _, infer __, infer TInternalState>
-    ? TInternalState
-    : never;
-
-export class Actor<
-  TLogic extends AnyActorLogic,
-  TEvent extends EventObject = EventFromLogic<TLogic>
-> implements ActorRef<TEvent, SnapshotFrom<TLogic>>
+export class Actor<TLogic extends AnyActorLogic>
+  implements ActorRef<EventFromLogic<TLogic>, SnapshotFrom<TLogic>>
 {
   /**
    * The current internal state of the actor.
    */
-  private _state!: InternalStateFrom<TLogic>;
+  private _state!: SnapshotFrom<TLogic>;
   /**
    * The clock that is responsible for setting and clearing timeouts, such as delayed events and transitions.
    */
@@ -101,7 +96,9 @@ export class Actor<
    */
   public id: string;
 
-  private mailbox: Mailbox<TEvent> = new Mailbox(this._process.bind(this));
+  private mailbox: Mailbox<EventFromLogic<TLogic>> = new Mailbox(
+    this._process.bind(this)
+  );
 
   private delayedEventsMap: Record<string, unknown> = {};
 
@@ -113,10 +110,14 @@ export class Actor<
   public status: ActorStatus = ActorStatus.NotStarted;
 
   // Actor Ref
-  public _parent?: ActorRef<any>;
-  public ref: ActorRef<TEvent>;
+  public _parent?: ActorRef<any, any>;
+  public ref: ActorRef<EventFromLogic<TLogic>, SnapshotFrom<TLogic>>;
   // TODO: add typings for system
-  private _actorContext: ActorContext<TEvent, SnapshotFrom<TLogic>, any>;
+  private _actorContext: ActorContext<
+    SnapshotFrom<TLogic>,
+    EventFromLogic<TLogic>,
+    any
+  >;
 
   private _systemId: string | undefined;
 
@@ -206,10 +207,9 @@ export class Actor<
   // array of functions to defer
   private _deferred: Array<() => void> = [];
 
-  private update(state: InternalStateFrom<TLogic>, event: EventObject): void {
+  private update(snapshot: SnapshotFrom<TLogic>, event: EventObject): void {
     // Update state
-    this._state = state;
-    const snapshot = this.getSnapshot();
+    this._state = snapshot;
 
     // Execute deferred effects
     let deferredFn: (typeof this._deferred)[number] | undefined;
@@ -227,21 +227,22 @@ export class Actor<
       }
     }
 
-    const status = this.logic.getStatus?.(state);
-
-    switch (status?.status) {
+    switch ((this._state as any).status) {
       case 'done':
         this._stopProcedure();
         this._complete();
-        this._doneEvent = createDoneActorEvent(this.id, status.data);
+        this._doneEvent = createDoneActorEvent(
+          this.id,
+          (this._state as any).output
+        );
         this.system.relay(this._doneEvent, this, this._parent);
 
         break;
       case 'error':
         this._stopProcedure();
-        this._error(status.data);
+        this._error((this._state as any).error);
         this.system.relay(
-          createErrorActorEvent(this.id, status.data),
+          createErrorActorEvent(this.id, (this._state as any).error),
           this,
           this._parent
         );
@@ -317,13 +318,16 @@ export class Actor<
       targetId: this.sessionId
     });
 
-    const status = this.logic.getStatus?.(this._state);
+    const status = (this._state as any).status;
 
-    switch (status?.status) {
+    switch (status) {
       case 'done':
         // a state machine can be "done" upon intialization (it could reach a final state using initial microsteps)
         // we still need to complete observers, flush deferreds etc
-        this.update(this._state, initEvent as unknown as TEvent);
+        this.update(
+          this._state,
+          initEvent as unknown as EventFromLogic<TLogic>
+        );
       // fallthrough
       case 'error':
         // TODO: rethink cleanup of observers, mailbox, etc
@@ -344,7 +348,7 @@ export class Actor<
     // TODO: this notifies all subscribers but usually this is redundant
     // there is no real change happening here
     // we need to rethink if this needs to be refactored
-    this.update(this._state, initEvent as unknown as TEvent);
+    this.update(this._state, initEvent as unknown as EventFromLogic<TLogic>);
 
     if (this.options.devTools) {
       this.attachDevTools();
@@ -355,7 +359,7 @@ export class Actor<
     return this;
   }
 
-  private _process(event: TEvent) {
+  private _process(event: EventFromLogic<TLogic>) {
     // TODO: reexamine what happens when an action (or a guard or smth) throws
     let nextState;
     let caughtError;
@@ -463,8 +467,16 @@ export class Actor<
     return this;
   }
 
-  /** @internal */
-  public _send(event: TEvent) {
+  /**
+   * @internal
+   */
+  public _send(event: EventFromLogic<TLogic>) {
+    if (typeof event === 'string') {
+      throw new Error(
+        `Only event objects may be sent to actors; use .send({ type: "${event}" }) instead`
+      );
+    }
+
     if (this.status === ActorStatus.Stopped) {
       // do nothing
       if (isDevelopment) {
@@ -485,7 +497,7 @@ export class Actor<
    *
    * @param event The event to send
    */
-  public send(event: TEvent) {
+  public send(event: EventFromLogic<TLogic>) {
     if (typeof event === 'string') {
       throw new Error(
         `Only event objects may be sent to actors; use .send({ type: "${event}" }) instead`
@@ -507,7 +519,7 @@ export class Actor<
     to?: AnyActorRef;
   }): void {
     const timerId = this.clock.setTimeout(() => {
-      this.system.relay(event as TEvent, this, to ?? this);
+      this.system.relay(event as EventFromLogic<TLogic>, this, to ?? this);
     }, delay);
 
     // TODO: consider the rehydration story here
@@ -546,9 +558,7 @@ export class Actor<
   }
 
   public getSnapshot(): SnapshotFrom<TLogic> {
-    return this.logic.getSnapshot
-      ? this.logic.getSnapshot(this._state)
-      : this._state;
+    return this._state;
   }
 }