+++ packages/core/test/actor.test.ts
@@ -1,28 +1,30 @@
-import {
-  createActor,
-  createMachine,
-  ActorRef,
-  ActorRefFrom,
-  EventObject,
-  ActorLogic,
-  Subscribable,
-  Observer,
-  AnyActorRef,
-  waitFor
-} from '../src/index.ts';
-import { sendParent, forwardTo } from '../src/actions.ts';
-import { raise } from '../src/actions/raise';
-import { assign } from '../src/actions/assign';
-import { sendTo } from '../src/actions/send';
 import { EMPTY, interval, of } from 'rxjs';
+import { map } from 'rxjs/operators';
+import { forwardTo, sendParent } from '../src/actions.ts';
+import { assign } from '../src/actions/assign';
+import { raise } from '../src/actions/raise';
+import { sendTo } from '../src/actions/send';
+import { CallbackActorRef, fromCallback } from '../src/actors/callback.ts';
+import {
+  fromEventObservable,
+  fromObservable
+} from '../src/actors/observable.ts';
+import { PromiseActorRef, fromPromise } from '../src/actors/promise.ts';
 import { fromTransition } from '../src/actors/transition.ts';
 import {
-  fromObservable,
-  fromEventObservable
-} from '../src/actors/observable.ts';
-import { fromPromise } from '../src/actors/promise.ts';
-import { fromCallback } from '../src/actors/callback.ts';
-import { map } from 'rxjs/operators';
+  ActorLogic,
+  ActorRef,
+  ActorRefFrom,
+  AnyActorRef,
+  EventObject,
+  Observer,
+  Snapshot,
+  Subscribable,
+  createActor,
+  createMachine,
+  stop,
+  waitFor
+} from '../src/index.ts';
 
 function sleep(ms: number) {
   return new Promise((res) => setTimeout(res, ms));
@@ -30,7 +32,7 @@ function sleep(ms: number) {
 
 describe('spawning machines', () => {
   const context = {
-    todoRefs: {} as Record<string, ActorRef<any>>
+    todoRefs: {} as Record<string, ActorRef<any, any>>
   };
 
   type TodoEvent =
@@ -74,7 +76,7 @@ describe('spawning machines', () => {
   });
 
   interface ClientContext {
-    server?: ActorRef<PingPongEvent>;
+    server?: ActorRef<PingPongEvent, Snapshot<unknown>>;
   }
 
   const clientMachine = createMachine({
@@ -235,7 +237,7 @@ describe('spawning promises', () => {
   it('should be able to spawn a promise', (done) => {
     const promiseMachine = createMachine({
       types: {} as {
-        context: { promiseRef?: ActorRef<any> };
+        context: { promiseRef?: PromiseActorRef<string> };
       },
       id: 'promise',
       initial: 'idle',
@@ -286,7 +288,7 @@ describe('spawning promises', () => {
     const promiseMachine = createMachine(
       {
         types: {} as {
-          context: { promiseRef?: ActorRef<any> };
+          context: { promiseRef?: PromiseActorRef<string> };
         },
         id: 'promise',
         initial: 'idle',
@@ -297,6 +299,7 @@ describe('spawning promises', () => {
           idle: {
             entry: assign({
               promiseRef: ({ spawn }) =>
+                // TODO: this is typed as AnyActorRef instead of PromiseActorRef<string>
                 spawn('somePromise', { id: 'my-promise' })
             }),
             on: {
@@ -338,7 +341,7 @@ describe('spawning callbacks', () => {
   it('should be able to spawn an actor from a callback', (done) => {
     const callbackMachine = createMachine({
       types: {} as {
-        context: { callbackRef?: ActorRef<any> };
+        context: { callbackRef?: CallbackActorRef<{ type: 'START' }> };
       },
       id: 'callback',
       initial: 'idle',
@@ -349,6 +352,7 @@ describe('spawning callbacks', () => {
         idle: {
           entry: assign({
             callbackRef: ({ spawn }) =>
+              // TODO: this is typed as AnyActorRef instead of CallbackActorRef<{ type: 'START' }>
               spawn(
                 fromCallback(({ sendBack, receive }) => {
                   receive((event) => {
@@ -409,7 +413,7 @@ describe('spawning observables', () => {
           on: {
             'xstate.snapshot.int': {
               target: 'success',
-              guard: ({ event }) => event.data === 5
+              guard: ({ event }) => event.data.context === 5
             }
           }
         },
@@ -445,7 +449,7 @@ describe('spawning observables', () => {
             on: {
               'xstate.snapshot.int': {
                 target: 'success',
-                guard: ({ event }) => event.data === 5
+                guard: ({ event }) => event.data.context === 5
               }
             }
           },
@@ -493,7 +497,8 @@ describe('spawning observables', () => {
               target: 'success',
               guard: ({ context, event }) => {
                 return (
-                  event.data === 1 && context.observableRef.getSnapshot() === 1
+                  event.data.context === 1 &&
+                  context.observableRef.getSnapshot().context === 1
                 );
               }
             }
@@ -536,7 +541,7 @@ describe('spawning observables', () => {
               onSnapshot: {
                 target: 'success',
                 guard: ({ event }) => {
-                  return event.data === 3;
+                  return event.data.context === 3;
                 }
               }
             }
@@ -561,10 +566,10 @@ describe('spawning observables', () => {
     const spy = jest.fn();
 
     actorRef.getSnapshot().children.childActor!.subscribe((data) => {
-      spy(data);
+      spy(data.context);
     });
 
-    await waitFor(actorRef, (state) => !!state.done);
+    await waitFor(actorRef, (state) => state.status !== 'active');
 
     expect(spy).toHaveBeenCalledWith(3);
   });
@@ -590,7 +595,7 @@ describe('spawning observables', () => {
               onSnapshot: {
                 target: 'success',
                 guard: ({ event }) => {
-                  return event.data === 3;
+                  return event.data.context === 3;
                 }
               }
             }
@@ -618,7 +623,7 @@ describe('spawning observables', () => {
       spy(data);
     });
 
-    await waitFor(actorRef, (state) => !!state.done);
+    await waitFor(actorRef, (state) => state.status !== 'active');
     spy.mockClear();
 
     // wait for potential next event from the interval actor
@@ -740,7 +745,7 @@ describe('communicating with spawned actors', () => {
 
     const parentMachine = createMachine({
       types: {} as {
-        context: { existingRef?: ActorRef<any> };
+        context: { existingRef?: typeof existingService };
       },
       initial: 'pending',
       context: {
@@ -805,7 +810,9 @@ describe('communicating with spawned actors', () => {
     const existingService = createActor(existingMachine).start();
 
     const parentMachine = createMachine({
-      types: {} as { context: { existingRef: ActorRef<any> | undefined } },
+      types: {} as {
+        context: { existingRef: typeof existingService | undefined };
+      },
       initial: 'pending',
       context: {
         existingRef: undefined
@@ -814,6 +821,7 @@ describe('communicating with spawned actors', () => {
         pending: {
           entry: assign({
             // TODO: fix (spawn existing service)
+            // @ts-expect-error
             existingRef: ({ spawn }) =>
               // @ts-expect-error
               spawn(existingService, {
@@ -887,7 +895,7 @@ describe('actors', () => {
     actor.start();
   });
 
-  it('should only spawn an actor in an initial state of a child that gets invoked in the initial state of a parent when the parent gets started', () => {
+  it('should spawn an actor in an initial state of a child that gets invoked in the initial state of a parent when the parent gets started', () => {
     let spawnCounter = 0;
 
     interface TestContext {
@@ -943,7 +951,7 @@ describe('actors', () => {
     });
 
     const testMachine = createMachine({
-      types: {} as { context: { ref?: ActorRef<any> } },
+      types: {} as { context: { ref?: ActorRefFrom<typeof anotherMachine> } },
       initial: 'testing',
       context: ({ spawn }) => {
         spawnCalled++;
@@ -971,7 +979,7 @@ describe('actors', () => {
 
   it('should spawn null actors if not used within a service', () => {
     const nullActorMachine = createMachine({
-      types: {} as { context: { ref?: ActorRef<any> } },
+      types: {} as { context: { ref?: PromiseActorRef<number> } },
       initial: 'foo',
       context: { ref: undefined },
       states: {
@@ -1067,12 +1075,13 @@ describe('actors', () => {
 
       const countService = createActor(countMachine);
       countService.subscribe((state) => {
-        if (state.context.count?.getSnapshot() === 2) {
+        if (state.context.count?.getSnapshot().context === 2) {
           done();
         }
       });
       countService.start();
 
+      debugger;
       countService.send({ type: 'INC' });
       countService.send({ type: 'INC' });
     });
@@ -1165,15 +1174,19 @@ describe('actors', () => {
     });
 
     it('actor logic should have reference to the parent', (done) => {
-      const pongLogic: ActorLogic<EventObject, undefined> = {
-        transition: (_state, event, { self }) => {
+      const pongLogic: ActorLogic<Snapshot<undefined>, EventObject> = {
+        transition: (state, event, { self }) => {
           if (event.type === 'PING') {
             self._parent?.send({ type: 'PONG' });
           }
 
-          return undefined;
+          return state;
         },
-        getInitialState: () => undefined
+        getInitialState: () => ({
+          status: 'active',
+          output: undefined,
+          error: undefined
+        })
       };
 
       const pingMachine = createMachine({
@@ -1216,9 +1229,10 @@ describe('actors', () => {
 
   it('should be able to spawn callback actors in (lazy) initial context', (done) => {
     const machine = createMachine({
-      types: {} as { context: { ref: ActorRef<any> } },
+      types: {} as { context: { ref: CallbackActorRef<{ type: 'TEST' }> } },
       context: ({ spawn }) => ({
         ref: spawn(
+          // TODO: this is typed as CallbackActorRef<EventObject> instead of CallbackActorRef<{ type: 'TEST' }>
           fromCallback(({ sendBack }) => {
             sendBack({ type: 'TEST' });
           })
@@ -1250,7 +1264,7 @@ describe('actors', () => {
     });
 
     const machine = createMachine({
-      types: {} as { context: { ref: ActorRef<any> } },
+      types: {} as { context: { ref: ActorRefFrom<typeof childMachine> } },
       context: ({ spawn }) => ({
         ref: spawn(childMachine)
       }),
@@ -1433,5 +1447,267 @@ describe('actors', () => {
 
     // Will be 2 if the event observable is resubscribed
     expect(subscriptionCount).toBe(1);
+  });
+
+  it('should be able to restart a spawned actor within a single macrostep', () => {
+    const actual: string[] = [];
+    let invokeCounter = 0;
+
+    const machine = createMachine({
+      types: {} as {
+        context: {
+          actorRef: CallbackActorRef<any>; // TODO: fix this
+        };
+      },
+      initial: 'active',
+      context: ({ spawn }) => {
+        const localId = ++invokeCounter;
+
+        return {
+          actorRef: spawn(
+            fromCallback(() => {
+              actual.push(`start ${localId}`);
+              return () => {
+                actual.push(`stop ${localId}`);
+              };
+            }),
+            { id: 'callback-1' }
+          )
+        };
+      },
+      states: {
+        active: {
+          on: {
+            update: {
+              actions: [
+                stop(({ context }) => {
+                  return context.actorRef;
+                }),
+                assign({
+                  actorRef: ({ spawn }) => {
+                    const localId = ++invokeCounter;
+
+                    return spawn(
+                      fromCallback(() => {
+                        actual.push(`start ${localId}`);
+                        return () => {
+                          actual.push(`stop ${localId}`);
+                        };
+                      }),
+                      { id: 'callback-2' }
+                    );
+                  }
+                })
+              ]
+            }
+          }
+        }
+      }
+    });
+
+    const service = createActor(machine).start();
+
+    actual.length = 0;
+
+    service.send({
+      type: 'update'
+    });
+
+    expect(actual).toEqual(['stop 1', 'start 2']);
+  });
+
+  it('should be able to restart a named spawned actor within a single macrostep when stopping by a ref', () => {
+    const actual: string[] = [];
+    let invokeCounter = 0;
+
+    const machine = createMachine({
+      types: {} as {
+        context: {
+          actorRef: CallbackActorRef<any>; // TODO: fix this
+        };
+      },
+      initial: 'active',
+      context: ({ spawn }) => {
+        const localId = ++invokeCounter;
+
+        return {
+          actorRef: spawn(
+            fromCallback(() => {
+              actual.push(`start ${localId}`);
+              return () => {
+                actual.push(`stop ${localId}`);
+              };
+            }),
+            { id: 'my_name' }
+          )
+        };
+      },
+      states: {
+        active: {
+          on: {
+            update: {
+              actions: [
+                stop(({ context }) => context.actorRef),
+                assign({
+                  actorRef: ({ spawn }) => {
+                    const localId = ++invokeCounter;
+
+                    return spawn(
+                      fromCallback(() => {
+                        actual.push(`start ${localId}`);
+                        return () => {
+                          actual.push(`stop ${localId}`);
+                        };
+                      }),
+                      { id: 'my_name' }
+                    );
+                  }
+                })
+              ]
+            }
+          }
+        }
+      }
+    });
+
+    const service = createActor(machine).start();
+
+    actual.length = 0;
+
+    service.send({
+      type: 'update'
+    });
+
+    expect(actual).toEqual(['stop 1', 'start 2']);
+  });
+
+  it('should be able to restart a named spawned actor within a single macrostep when stopping by static name', () => {
+    const actual: string[] = [];
+    let invokeCounter = 0;
+
+    const machine = createMachine({
+      types: {} as {
+        context: {
+          actorRef: CallbackActorRef<never>;
+        };
+      },
+      initial: 'active',
+      context: ({ spawn }) => {
+        const localId = ++invokeCounter;
+
+        return {
+          actorRef: spawn(
+            fromCallback(() => {
+              actual.push(`start ${localId}`);
+              return () => {
+                actual.push(`stop ${localId}`);
+              };
+            }),
+            { id: 'my_name' }
+          )
+        };
+      },
+      states: {
+        active: {
+          on: {
+            update: {
+              actions: [
+                stop('my_name'),
+                assign({
+                  actorRef: ({ spawn }) => {
+                    const localId = ++invokeCounter;
+
+                    return spawn(
+                      fromCallback(() => {
+                        actual.push(`start ${localId}`);
+                        return () => {
+                          actual.push(`stop ${localId}`);
+                        };
+                      }),
+                      { id: 'my_name' }
+                    );
+                  }
+                })
+              ]
+            }
+          }
+        }
+      }
+    });
+
+    const service = createActor(machine).start();
+
+    actual.length = 0;
+
+    service.send({
+      type: 'update'
+    });
+
+    expect(actual).toEqual(['stop 1', 'start 2']);
+  });
+
+  it('should be able to restart a named spawned actor within a single macrostep when stopping by resolved name', () => {
+    const actual: string[] = [];
+    let invokeCounter = 0;
+
+    const machine = createMachine({
+      types: {} as {
+        context: {
+          actorRef: CallbackActorRef<never>;
+        };
+      },
+      initial: 'active',
+      context: ({ spawn }) => {
+        const localId = ++invokeCounter;
+        actual.push(`start ${localId}`);
+
+        return {
+          actorRef: spawn(
+            fromCallback(() => {
+              return () => {
+                actual.push(`stop ${localId}`);
+              };
+            }),
+            { id: 'my_name' }
+          )
+        };
+      },
+      states: {
+        active: {
+          on: {
+            update: {
+              actions: [
+                stop(() => 'my_name'),
+                assign({
+                  actorRef: ({ spawn }) => {
+                    const localId = ++invokeCounter;
+
+                    return spawn(
+                      fromCallback(() => {
+                        actual.push(`start ${localId}`);
+                        return () => {
+                          actual.push(`stop ${localId}`);
+                        };
+                      }),
+                      { id: 'my_name' }
+                    );
+                  }
+                })
+              ]
+            }
+          }
+        }
+      }
+    });
+
+    const service = createActor(machine).start();
+
+    actual.length = 0;
+
+    service.send({
+      type: 'update'
+    });
+
+    expect(actual).toEqual(['stop 1', 'start 2']);
   });
 });