+++ packages/core/test/after.test.ts
@@ -213,7 +213,9 @@ describe('delayed transitions', () => {
 
     const persistedState = JSON.parse(JSON.stringify(service.getSnapshot()));
 
-    service = createActor(createMyMachine(), { state: persistedState }).start();
+    service = createActor(createMyMachine(), {
+      state: persistedState
+    }).start();
 
     service.send({ type: 'NEXT' });