+++ packages/core/test/invoke.test.ts
@@ -19,7 +19,8 @@ import {
   createMachine,
   createActor,
   sendParent,
-  EventFrom
+  EventFrom,
+  Snapshot
 } from '../src/index.ts';
 
 const user = { name: 'David' };
@@ -585,7 +586,7 @@ describe('invoke', () => {
       actor.start();
     });
 
-    it('should not reinvoke root-level invocations', (done) => {
+    it('should not reinvoke root-level invocations on root non-reentering transitions', () => {
       // https://github.com/statelyai/xstate/issues/2147
 
       let invokeCount = 0;
@@ -630,7 +631,6 @@ describe('invoke', () => {
       expect(invokeCount).toEqual(1);
       expect(invokeDisposeCount).toEqual(0);
       expect(actionsCount).toEqual(2);
-      done();
     });
 
     it('should stop a child actor when reaching a final state', () => {
@@ -870,7 +870,7 @@ describe('invoke', () => {
         const service = createActor(promiseMachine);
         service.subscribe({
           error(err) {
-            expect(err.message).toEqual(expect.stringMatching(/test/));
+            expect((err as any).message).toEqual(expect.stringMatching(/test/));
             done();
           }
         });
@@ -906,7 +906,7 @@ describe('invoke', () => {
         actor.subscribe({
           error: (err) => {
             expect(err).toBeInstanceOf(Error);
-            expect(err.message).toBe('test');
+            expect((err as any).message).toBe('test');
             expect(completeSpy).not.toHaveBeenCalled();
             done();
           },
@@ -1879,7 +1879,7 @@ describe('invoke', () => {
             invoke: {
               src: fromObservable(() => interval(10)),
               onSnapshot: {
-                actions: assign({ count: ({ event }) => event.data })
+                actions: assign({ count: ({ event }) => event.data.context })
               }
             },
             always: {
@@ -1923,7 +1923,7 @@ describe('invoke', () => {
               src: fromObservable(() => interval(10).pipe(take(5))),
               onSnapshot: {
                 actions: assign({
-                  count: ({ event }) => event.data
+                  count: ({ event }) => event.data.context
                 })
               },
               onDone: {
@@ -1975,7 +1975,7 @@ describe('invoke', () => {
                 )
               ),
               onSnapshot: {
-                actions: assign({ count: ({ event }) => event.data })
+                actions: assign({ count: ({ event }) => event.data.context })
               },
               onError: {
                 target: 'success',
@@ -2011,7 +2011,7 @@ describe('invoke', () => {
           input: 42,
           onSnapshot: {
             actions: ({ event }) => {
-              expect(event.data).toEqual(42);
+              expect(event.data.context).toEqual(42);
               done();
             }
           }
@@ -2203,16 +2203,30 @@ describe('invoke', () => {
 
   describe('with logic', () => {
     it('should work with actor logic', (done) => {
-      const countLogic: ActorLogic<EventObject, number> = {
-        transition: (count, event) => {
+      const countLogic: ActorLogic<
+        Snapshot<undefined> & { context: number },
+        EventObject
+      > = {
+        transition: (state, event) => {
           if (event.type === 'INC') {
-            return count + 1;
+            return {
+              ...state,
+              context: state.context + 1
+            };
           } else if (event.type === 'DEC') {
-            return count - 1;
+            return {
+              ...state,
+              context: state.context - 1
+            };
           }
-          return count;
+          return state;
         },
-        getInitialState: () => 0
+        getInitialState: () => ({
+          status: 'active',
+          output: undefined,
+          error: undefined,
+          context: 0
+        })
       };
 
       const countMachine = createMachine({
@@ -2229,7 +2243,7 @@ describe('invoke', () => {
 
       const countService = createActor(countMachine);
       countService.subscribe((state) => {
-        if (state.children['count']?.getSnapshot() === 2) {
+        if (state.children['count']?.getSnapshot().context === 2) {
           done();
         }
       });
@@ -2240,15 +2254,19 @@ describe('invoke', () => {
     });
 
     it('logic should have reference to the parent', (done) => {
-      const pongLogic: ActorLogic<EventObject, undefined> = {
-        transition: (_, event, { self }) => {
+      const pongLogic: ActorLogic<Snapshot<undefined>, EventObject> = {
+        transition: (state, event, { self }) => {
           if (event.type === 'PING') {
             self._parent?.send({ type: 'PONG' });
           }
 
-          return undefined;
+          return state;
         },
-        getInitialState: () => undefined
+        getInitialState: () => ({
+          status: 'active',
+          output: undefined,
+          error: undefined
+        })
       };
 
       const pingMachine = createMachine({
@@ -2308,7 +2326,7 @@ describe('invoke', () => {
 
       const countService = createActor(countMachine);
       countService.subscribe((state) => {
-        if (state.children['count']?.getSnapshot() === 2) {
+        if (state.children['count']?.getSnapshot().context === 2) {
           done();
         }
       });
@@ -2324,7 +2342,7 @@ describe('invoke', () => {
       const countReducer = (
         count: number,
         event: CountEvents,
-        { self }: ActorContext<CountEvents, any>
+        { self }: ActorContext<any, CountEvents>
       ): number => {
         if (event.type === 'INC') {
           self.send({ type: 'DOUBLE' });
@@ -2351,7 +2369,7 @@ describe('invoke', () => {
 
       const countService = createActor(countMachine);
       countService.subscribe((state) => {
-        if (state.children['count']?.getSnapshot() === 2) {
+        if (state.children['count']?.getSnapshot().context === 2) {
           done();
         }
       });
@@ -3221,6 +3239,51 @@ describe('invoke', () => {
     service.send({ type: 'EVENT' });
 
     expect(count).toEqual(2);
+  });
+
+  it('should be able to restart an invoke when reentering the invoking state', () => {
+    const actual: string[] = [];
+    let invokeCounter = 0;
+
+    const machine = createMachine({
+      initial: 'inactive',
+      states: {
+        inactive: {
+          on: { ACTIVATE: 'active' }
+        },
+        active: {
+          invoke: {
+            src: fromCallback(() => {
+              const localId = ++invokeCounter;
+              actual.push(`start ${localId}`);
+              return () => {
+                actual.push(`stop ${localId}`);
+              };
+            })
+          },
+          on: {
+            REENTER: {
+              target: 'active',
+              reenter: true
+            }
+          }
+        }
+      }
+    });
+
+    const service = createActor(machine).start();
+
+    service.send({
+      type: 'ACTIVATE'
+    });
+
+    actual.length = 0;
+
+    service.send({
+      type: 'REENTER'
+    });
+
+    expect(actual).toEqual(['stop 1', 'start 2']);
   });
 });