+++ packages/xstate-solid/src/types.ts
@@ -7,25 +7,33 @@ import type {
   ActorOptions,
   MachineContext,
   ProvidedActor,
-  State,
-  TypegenDisabled
+  TypegenDisabled,
+  HomomorphicPick,
+  MachineSnapshot
 } from 'xstate';
 
-type StateObject<
+type MachineSnapshotPOJO<
   TContext extends MachineContext,
   TEvent extends EventObject = EventObject,
   TActor extends ProvidedActor = ProvidedActor,
   TTag extends string = string,
   TOutput = unknown,
   TResolvedTypesMeta = TypegenDisabled
-> = Pick<
-  State<TContext, TEvent, TActor, TTag, TOutput, TResolvedTypesMeta>,
-  keyof AnyState
+> = HomomorphicPick<
+  MachineSnapshot<TContext, TEvent, TActor, TTag, TOutput, TResolvedTypesMeta>,
+  keyof MachineSnapshot<
+    TContext,
+    TEvent,
+    TActor,
+    TTag,
+    TOutput,
+    TResolvedTypesMeta
+  >
 >;
 
 // Converts a State class type to a POJO State type. This reflects that the state
 // is being spread into a new object for reactive tracking in SolidJS
-export type CheckSnapshot<Snapshot> = Snapshot extends State<
+export type CheckSnapshot<Snapshot> = Snapshot extends MachineSnapshot<
   infer TContext,
   infer TEvents,
   infer TActor,
@@ -33,7 +41,14 @@ export type CheckSnapshot<Snapshot> = Snapshot extends State<
   infer TOutput,
   infer TResolvedTypesMeta
 >
-  ? StateObject<TContext, TEvents, TActor, TTag, TOutput, TResolvedTypesMeta>
+  ? MachineSnapshotPOJO<
+      TContext,
+      TEvents,
+      TActor,
+      TTag,
+      TOutput,
+      TResolvedTypesMeta
+    >
   : Snapshot;
 
 type InternalMachineOpts<