+++ packages/xstate-react/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 4.0.0-beta.10
+
+### Minor Changes
+
+- [#4240](https://github.com/statelyai/xstate/pull/4240) [`409552cf8`](https://github.com/statelyai/xstate/commit/409552cf8578dc517cca4e6eaeb652a33f49c06c) Thanks [@davidkpiano](https://github.com/davidkpiano)! - The `useMachine` function is aliased to `useActor` and not shown as visually deprecated.
+
 ## 4.0.0-beta.9
 
 ### Major Changes